api/requirements.txt
aiofiles==23.1.0
aiogTTS==1.1.1
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==2.0.1
aiosqlite==0.16.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
autopep8==2.0.2
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
cffi==1.14.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fakeredis==1.10.2
fastapi==0.95.1
fastapi-mail==1.2.7
frozenlist==1.3.3
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
priority==1.3.0
pycodestyle==2.10.0
pycparser==2.20
pydantic==1.9.2
PyJWT==2.0.0
PyPika==0.44.1
pypika-tortoise==0.1.6
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2020.5
redis==4.4.4
requests==2.28.2
rfc3986==1.5.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.26.1
toml==0.10.2
tomli==2.0.1
tortoise-orm==0.19.3
typing-extensions==3.7.4.3
urllib3==1.26.15
uvicorn==0.21.1
wsproto==1.0.0
yarl==1.8.2
